% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Sum two values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{An integer or NA.}

\item{y}{An integer or NA. If not specified, the default value is 10.}
}
\value{
An integer sum or NA.
}
\description{
Sum two values
}
\examples{
my_add(1, 2)
my_add(1)
my_add(NA)
my_add(10, NA)

}
